Interpolation and Expressions
Showing data dynamically to the user on the DOM via interpolation. Message is initialised on ngOnInit instead of on constructor as when component gets mounted, the ngOnInit gets called after the constructor is instantiated, which is a hook that Angular invokes. Once the component is bootstrapped by Angular, then it will run ngOnInit lifecycle hook and assign the this.message value. The ! in 'message!:' tells typescript that the property is gonna be assigned and will be available later on (which is on ngOnInit). Added ternary statement to say that if there is a message, print out the message and if there isn't one, then print 'Nothing here...'