Property Binding Syntax
Showcasing how the double curly braces interpolation syntax is simply sugar syntax, which is a nicer way to display values without binding to a property directly. And the way Angular does this under the hood is the same as the innerText property binding. Also binding the value of message to the property of value to show how to use one way data binding syntax correctly.