Event Binding Syntax
Binding a click event with the handleClick function. The function itself just logs the event object (PointerEvent) which provides lots of properties in the DOM. Binding a input event with the handleInput function. Destructuring the value within the function, whilst telling typescript that the event.target is being type casted as a HTMLInputElement. With newMessage the way that works is that the data is taken from message, then it gets changed from the data that is taken from the change event. Then we set it to the new variable.

side note:
In terms of component communication. The way data is passed through is via property binding '[value]' and the way we get data from component is via event binding '(input)'.