Template Ref Variables
This concept allows us to grab a hold of an element or component and access some further information about it. The # gives us a variable that we can access throughout the template and pluck pieces of information from. The messageInput value is exported to the interpolated <p> tag. Within the event binding, the messageInput value is reassigned to the newMessage value. This is showcasing a simplistic approach without using callback functions.