Create a Child Feature Module
Building out first child feature module using angular CLI via command "npx ng generate module admin --module=app.module". This is to understand the flow of creating a base level component, which will end up being used for routing purposes. The purpose of the admin module is to allow us to manage our products. So instead of adding logic to the app module, the logic will be added to the admin module and then simply import it. This is so that if we wanted to used the admin module in another project, we can simply take the whole folder and with one import, it can be imported to another application.