Create a Smart Container Component
Building a way for to manage the products for a donut shop via
"npx ng generate component admin/containers/donut-list" command. This is a smart component which will be in charge of fetching the data and handing it down to the child components. Important thing to note when declaring the <donut-list></donut-list> component in the app component will error of "donut-list is not a known element". To resolve this need to export the component inside the admin module to allow it to be used elsewhere.