Use the NgOnInit Lifecycle Hook
Started creating some template data and rendering it out in the DOM using the NgOnInit lifecycle hook. The idea of having the data within inner <div> is that this is a container component, this component is going to contain sub components (child components). The array of data will be passed down and render all of those components. For now the data will be built out, then break it down into smaller pieces. Then work out as we go how the flow of things actually works.