Create and Render a Dumb Component
Creating the template for the individual donut card, then converting it to it's own reusable dumb component. Which we can then pass data down to these dumb components for one way data flow. Using the donut.icon as a slug for getting the donut svgs. Created the dumb component via the "npx ng generate component admin/components/donut-card" command. Will fix the issue of the properties not existing on the dumb component in the next commit.