Pass State to Dumb Components via @Input
Adding a donut property to the <donut-card> which will have a property binding which will have the value of "donut", this will be the data that is going to be passed down. Using the Input to notify Angular that donut is an input binding (data comes into the component via an input). The @Input() typescript decorator is doing some Angular magic on top of the donut property to tell Angular that it is expecting some data to be bound from the parent (donut-list component). Duplicating the <donut-card> components with different hardcoded values from the donuts array to highlight how easy it is to reuse that particular component. Will work on how to control the flow of the application and render these out dynamically based on a list.