Style a Host Element
Demonstrating styling a host element such as the <donut-card> selector within the donut-list component. This method helps to avoid having to avoid styling the component without having to repeatedly write the styles in the parent component each time. The host is part of the shadow DOM specification which is a newer feature of the web platform.