Set Inline Styles with Style Bindings
Binding a specific inline style (style attribute) within the template. Added a new optional property called "promo" which is a concept of a product that is brand new. Setting an inline style based off of whether if a product is in promotion. Using a nicer syntax to bind the style to the property of border itself rather than do:
style = "border: ..."
learned how to conditionally add some styles. Looked at the style property and bound to a border. This can also be thought of as plain javascript similar to when running this command into the browser console:
document.body.style.border = '5px solid red"