Complex Inline Styles with NgStyle
Using first directive which is a way we can bind specific behaviour to an element or a component. In this case we are binding to an element via ngStyle directive. Which is a more super-powered style binding which allows us to specify an object, and get a bit more complex when we need to set multiple inline styles. Another style that can be demostrated would be adding:
'font-size.px' : 20
Which assumes that we get data back from a server of the number 20. This method allows for specifying the unit as an extension of the property we want to use which in this case is pixels.