Set Classes with Class Bindings
Understanding how to add classes conditionally. Whilst inline styles are a valid use case, they don't allow us to manage styles effectively because it limits our re-use. By using classes we can move the style into a class selector, and means we don't need to add styles that don't need adding. By this it means that we don't need to use 'none' to specify adding no styles. Adding the class conditionally we can simply remove it. When using [class.donut-card-promo] this is not setting an empty class, it is actually removing it altogether (like a toggle), which is based off the truthy expression of "donut.promo".