Use NgIf for Conditional Rendering
Rendering flow of components within the app. We have rendering where Angular can render a particular component. And we have flow of rendering the component. Can also have Angular interpret our component templating, and combine the two in terms of a rendering and flow of rendering. Using ngIf statement to the template which is a directive. The asterisk is a way to declare part of the template to perform a certain behaviour. This has introduced the concept of an empty container div which will be worked on in the next commit.