Use <ng-container> for Virtual Elements
With the child div where the asterisk means a template, it could pose layout issues in the future when the app grows. For example if we wanted to use flexbox on the parent div, we are unable to directly target the nodes (donut-card elements) inside. We can use virtual element (ng-container) because it does not get rendered in the DOM. Ng-container was added as a bit more flexible container where we can use any directives. A nice way to jump out of the rendered HTML world and just simply control the components that are then rendered below.