Use the NgIf, Else and Then Syntax
Creating portions of the template if there are donut cards, else display the message if there is no donut cards. The no donuts message would be typically seen if we are waiting for a particular request to come back. This is a nice way we can break out the components into different templates inside the same component, this then gives us a single place we can perform some complex operations on which cleans up the code.