NgIf Syntax with <ng-template>
Showcasing how the ng-template doesn't need the asterisk for the ngIf because the asterisk is an shorthand sugar syntax for creating this ng-template. The <ng-container *ngIf... is exactly the same as <ng-template [ngIf]..., but the main difference between the two is that the ng-container will render things out without the *ngIf. Whereas removing the [ngIf] from ng-template would not render anything. By using ng-template, we're creating a template to then be re-used at a future point in time, whereas ng-container is simply for grouping elements.