NgFor and Iterating Collections
Implemented a javascript style of a For...Of loop by using NgFor which is a dedicated directive for simply repeating an element based on the data we give it. The way "let donut of donuts" works is the "donuts" is a iterable collection. The "let donut" is passed in each single item so that it displays the items on the DOM.