Improve NgFor Performance with TrackBy
Using Angulars in built in change detection. Every cycle of this change detection that gets triggered, for our For...Of loop, Angular uses an internal differ. This diffing process is simply comparing by identity (identity of objects). As we're building with an immutable approach with one way data flow, this is gonna cause a rendering problem, in that Angular is going to compare these objects that we are given back through donuts because the identity has changed each time. The trackById function works so that if an item is deleted from the data set, we are fully optimising how Angular is performing it's diff checking and the re-rendering of the components. This is a simply quick one line performance improvement that we can pass to Angulars ngFor directive and we want to say "I want to trackBy..." and provide the function name of the implementation of how we want Angular to then go ahead and start tracking our changes. Angular can very quickly compare the IDs to the last set of IDs, and then go ahead and re-render our components extremely quickly.