Use NgSwitch for Case Rendering
Using ngSwitch which is basically a switch statement in Javascript turned into an Angular directive. Changed the promo property to take in certain strings to allow us to change the label next to the title of a particular donut. This is a great way to supply some elements to easily be conditionally rendered based on the data we are passing.