Advanced NgSwitch Syntax with <ng-template>
Implemented a slightly more advanced use case by using the ng-template syntax. From using ng-container and ng-template it means that it allows us to specify the word "NEW" without a containing element (basically they do not get rendered). Reason why we might want to use this method because the code is presented in a simpler fashion. Also if you have a piece of DOM (<span class="donut-card-label">) that is used in every single scenario, this might be a great way to handle things.