Create a Form Container and Dumb Component
Hooking up the ability to create a definition for a new product. Can consist of creating a name, adding price, save things out and pass this data eventually off to a server which is then save things in our local database. Creating two new components, first is a container component (npx ng generate component admin/containers/donut-single) which will eventually going to tie up to some routing. The second being a dumb  form component (ng generate component admin/components/donut-form). which will allow us to create a new product, and then add this to our data structure.
In more detail the <donut-single> is in charge of communicating with the services hence why we are creating a container.
The <donut-form> will have some events such as a create, save, update and delete event