Use NgForm with a Template Ref Variable
Started to build out the donut form. The #form="ngForm" template variable is used to register our form with Angular. This essentially register our form as an Angular form, so Angular behind the scenes is going to look for different directives and start to build a form model for us in the background. As well as keep track of changes and lots of different things.
The form value is wrapped in the pre tag to allow us to have some nicer formatting as we build out our form.