Radio Button Binding with NgModel
Binding values to a radio button via ngModel, as well as unbind them. Used property binding for the None radio button to pass in Javascript of undefined. This is because this causes the promo property to be obliterated from the object altogther. This helps to submit the form, post it off and not add pointless properties that are never used to our data structure payload.