Show Error Messages on Validation Change
Implemented validation such as "required" and "minLength" which are tied in to the ng-container. The name.invalid and name.touched properties means that at this point we have an error. This is where the name.errors?.minLength or name.errors?.required are used to find the particular error which is given to us one by one to display the relevant styled HTML text.