Style Invalid Input Fields Globally
The global styling was already implemented so adding a space to the comment to highlight the below styling code. The "form .ng-invalid" indicates that it's looking for any child elements or components that contain the ng-valid inside the form. If it was chained "form.ng-invalid" then the behaviour would be different in that we are only going to see the ng-valid class states being applied to the form itself. The styling will only get painted if our element contains both .ng-invalid.ng-touched at the same time. Adding it into the styles.scss files because we might end up having another form, so adding it here makes sense to save on duplicating code.