Disable Submit and Safety Check Submissions
Added a disabled property binding when the form is invalid which causes the button to be disabled which prevents the user to submit invalid data. Also added a final safety check in the form of an if check in the handleSubmit function to only return the form value when the form is valid. This safety measure prevents any invalid data being passed through if someone were to tamper with the DOM (deleting the disabled tag).