Control NgModel Writes with ngModelOptions
Updating our model at a different point in time than when we are typing. This is in the event of a new requirement where we are typing in an input and don't want to keep searching on the backend (auto-complete feature). The idea behind "{ updateOn: 'blur' }" is so that when a user is typing something (focus event), the form.value property doesn't get updated in realtime, and only updates itself when the user clicks off the input field (fires a blur event).