Trigger a Form State and Validation Reset
Implemented a feature to reset the form. Binded a click event to the reset form button, at first used form.reset() which is aimed at a form control. Whereas form.resetForm() is aimed at the form, it's validation and all things with it (such as form.submitted event). If we wanted a true 100% resetted form with all the flags underneath, it is preferred to use form.resetForm().