Share State in Components via Services
Moving to a database driven application, firstly by moving the data from the component and putting it into a state service. Added logic in the single component to find the item in the service and renders it if it matches the id. Otherwise render the fallback if the id does not match what is in the service (creating a new donut).