Class Methods as State Selectors
Embracing idea of state selector in that we are selecting a piece of our state based on an id. Also exploring idea of class method as a state selector to encapsulate our app state, which then corresponds to certain features in the app, and wrap them up nicely and hide the implementation inside of a class method. Using pattern of CRUD.