Immutable State Updates
Looked at the data structure, created a update method which then immutatbly updates our data structure which returns a new array every single time. Then re-assigns that value to the this.donuts. This is then shared because we are accessing it inside of our donut service. We created that nice wrapper where we are selecting a particular item (via id). We are using onCreate and onUpdate which are two different events.