Read Data with HttpClient GET
Implemented GET request on read function so that we can populate state property inside of the service. This means that the service for 2 things such as holding state and making API requests. Imported Angulars HTTP client module to make the API calls in the service. In terms of using RXJS the use of subscribe, it simply means giving out the data when it is ready.