Store State with Observable Pipes
Cached data by taking the response from the http GET and cache it by essentially putting it inside the Donut array. Used a pipe which allows us to specify a few things which are going to happen after the HTTP GET has been completed. Within the pipe we are going to tap into the response which allows us to provide a side effect after the HTTP has been completed and give us the donut. The benefit of doing this is that we can do an if check to see there is a donuts.length, then we know that the data is already loaded in which case we can simply return this.donuts. This in turn allows us to then specify some nice caching and not making another HTTP request. The of(this.donuts) which all it is doing is allowing us to create own observable, and it doesn't matter what data we pass into of. Essentially using an observable wrapper around our data set, the only reason we do this is so that we can then inside of our list call subscribe directly. So what this means is if we don't have any data, then our read function is going to take care of the HTTP GET, then we can subscribe to the returned observable and get the data. Or if the data is already available, we are just going to pass it back as an observable anyway.

Fixing an console error of "Cannot read properties of undefined (reading 'length')" within donut list component. This is due to the fact that our donuts has no initial value.