Branch off an Observable to Access State
Implemented the ability to fetch the data for our donut single component. Added a subscribe after the readOne function in the donut single component which which allows us to do is get that individual donut, and then bind it to this.donut. Within the readOne function in the service, we pass in the map function so that when we map over the items, we are then going to get given all of the donuts. So inside of our callback is for us to implement our find. If we do find the donut then we want to return it, otherwise we are going to return an empty object.

Added a loading display as the data gets returned to us after 1 second so that user experience wise it feels far nicer.