Create Data with HttpClient POST
Implementing adding a new donut to the database by putting something that doesn't exist and saving it out.
Within the single component the OnCreate call needs the subscribe otherwise it is going to do nothing. Observables by design are lazy and only trigger when we subscribe to them and attach data. This will have a function be passed in via the console log.
Within the service we call the create, we pass the payload in, Angulars Http POST kicks off. Goes to the API, passed the payload off. Then it gets saved in the database which the ID gets added. We then can jump into the observable using TAP. We know that it is a single donut because we provided the type definitions here. Then using our existing logic to set the state and add that item to our application state. We then tack it on the end of our existing array.