Update Data with HttpClient PUT
Within the service the update method will have a PUT request to be made, we pass in the ID of the item we want to update. We then supply the item in it's updated form. We then break out before our response gets returned, to our single component. We then get given that individual updated donut property back from the server. We then reassign the value of our donuts. Because what will happen is that after we hit update, we are going back to the homepage and see those changes being reflected instantly. They are instant because we are updating our state, which is then going to be shared amongst our components. So the updated state will be the result of calling donuts, we then using the MAP to supply a new array. Any items we are going to pass back that do not match the ID of the item which was then updated. We are just making a quick swap in our data structure based on the item that's come back