Use retryWhen() to Delay Retries
Learnt how to successfully retry a failed HTTP error. Merged things over to retryWhen (even though its been deprecated), learned about errors which is given to us as an observable, then pipe which allows us to pass in any operators that we like. Dealing with pure functions (delay and take) which then pass the values to the next. When our delay fires, our take is going to kick off, it's going to take the first value, then the second value which is then going to finish that observable. But if we come back online beforehand. It's not going to take it twice because we are not going to have a retry if it was successful the first time.

Alternative to retryWhen from someone also doing the same course:
https://github.com/ultimatecourses/angular-basics-seed/issues/3