Redirect Routes to New Paths
Using full path match when path is empty. Supplied a nice quick easy redirect. As well as relative redirects, where once we are routed to a component (redirectTo admin), then go inside a particular path (path admin), it is the up to the child route to then take over and then redirect.