Pass Custom Data to Routed Components
Taking admin modules data property and passing "isEdit" to it. Using the existing activated route that was imported, and created new property in smart component. Took "isEdit" property and bound it to the donut form. Pass it down to the form and set up functionality to show/hide buttons accordingly. Lastly made a small feature edition that only toggling the "reset form" button, if the user has actually edited it. Otherwise they can see it when they hit up an existing item.