Providing a Route Config via importProvidersFro
Covered using "importProvidersFrom" the RouterModule to allow us then inject some of the built in Angular modules because the RouterModule is not a component. This will also apply to some other built in Angular modules. This is a standalone approach for the time being and how to achieve a fully standalone application. Refactored "app.module" into a "app.routes". Have a very nice and clean "app.routes" file with just a routing definition and a nice const that is exported, and nothing else.