Lazy-Load Component-Less Route Configs
With the router having some configuration changes which allows us to easily load the routing definitions by just passing in this loadChildren. Which is nice because inside of the "admin.routes", everything is then scoped. This allows us to load our entire routing definition from another module in a standalone fashion