Providing Standalone Directives, Components and Pipes
Converted application over to a standalone approach. Imported the directives and components individually. Converted each of these components into a standalone component. Ended up with a "admin.routes" file which has the components and other stuff that we no longer need to specify. Left with a nice file with container components that are being used as routing definitions.